langchain/day2/requirements.txt
langchain 
langchain-community 
langchain-ollama
python-dotenv
redis